+++ samples/test/logs.test.js
@@ -15,21 +15,19 @@
 'use strict';
 
 const {assert} = require('chai');
-const {describe, it} = require('mocha');
+const {describe, it, after} = require('mocha');
 const cp = require('child_process');
 const uuid = require('uuid');
 
 const execSync = cmd => cp.execSync(cmd, {encoding: 'utf-8'});
 const cmd = 'node logs';
-const logName = `nodejs-logging-logs-test-${uuid.v4()}`;
-const message = 'Hello world!';
+const TESTS_PREFIX = 'nodejs-docs-samples-test';
+const logName = `${TESTS_PREFIX}-${Date.now()}-${uuid.v4().split('-').pop()}`;
 
 describe('logs', () => {
-  it('should write a log entry', () => {
-    const output = execSync(
-      `${cmd} write ${logName} '{"type":"global"}' '{"message":"${message}"}'`
-    );
-    assert.include(output, `Wrote to ${logName}`);
+  after(async () => {
+    const oneHourAgo = new Date();
+    oneHourAgo.setHours(oneHourAgo.getHours() - 1);
   });
 
   it('should write a simple log entry', () => {